+++ app/src/main/java/com/hollowvyn/kneatr/data/local/entity/TagEntity.kt
@@ -5,6 +5,6 @@ import androidx.room.PrimaryKey
 
 @Entity(tableName = "tags")
 data class TagEntity(
-    @PrimaryKey(autoGenerate = true) val tagId: Int,
+    @PrimaryKey(autoGenerate = true) val tagId: Long,
     val name: String,
 )
